Move `preprocess` functions and tests new `clean` and `commongrid` subpackages (#993)
* Move noise functions and tests from preprocess to new filter subpackage

* Fix docstring for new filter/api.py module

* Add missing estimate_noise to filter __init__

* Move remaining preprocess modules and test to commongrid; update all references to preprocess

* Update provenance attribute strings to replace preprocess with commongrid or filter

* Create stub preprocess subpackage so function calls still work, with a deprecation warning

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Rename new 'filter' subpackage to 'clean', to avoid Python reserved word

* Update echopype/clean/api.py


* Update .ci_helpers/run-test.py

* Update echopype/commongrid/api.py


* Update echopype/__init__.py

Sorting __all__ strings in alphabetical order

---------
