+++ .ci_helpers/run-test.py
@@ -25,15 +25,17 @@ EXIT_CODES = {
 
 MODULES_TO_TEST = {
     "root": {},  # This is to test the root folder.
-    "convert": {},
     "calibrate": {},
+    "clean": {},
+    "commongrid": {},
+    "consolidate": {},
+    "convert": {},
     "echodata": {},
+    "mask": {},
+    "metrics": {},
     "preprocess": {},
     "utils": {},
     "visualize": {},
-    "metrics": {},
-    "mask": {},
-    "consolidate": {},
 }
 
 if __name__ == "__main__":
