+++ echopype/__init__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 from _echopype_version import version as __version__  # noqa
 
-from . import calibrate, consolidate, mask, preprocess, utils
+from . import calibrate, clean, commongrid, consolidate, mask, preprocess, utils
 from .convert.api import open_raw
 from .echodata.api import open_converted
 from .echodata.combine import combine_echodata
@@ -15,12 +15,14 @@ verbose(override=True)
 init_ep_dir()
 
 __all__ = [
-    "open_raw",
-    "open_converted",
-    "combine_echodata",
     "calibrate",
+    "clean",
+    "combine_echodata",
+    "commongrid",
     "consolidate",
     "mask",
+    "open_converted",
+    "open_raw",
     "preprocess",
     "utils",
     "verbose",
