+++ echopype/clean/__init__.py
@@ -1,0 +1,6 @@
+from .api import estimate_noise, remove_noise
+
+__all__ = [
+    "estimate_noise",
+    "remove_noise",
+]
