+++ echopype/commongrid/__init__.py
@@ -1,0 +1,6 @@
+from .api import compute_MVBS, compute_MVBS_index_binning
+
+__all__ = [
+    "compute_MVBS",
+    "compute_MVBS_index_binning",
+]
