+++ echopype/consolidate/api.py
@@ -296,7 +296,7 @@ def add_splitbeam_angle(
         source_Sv = xr.open_dataset(source_Sv, engine=file_type, chunks={}, **storage_options)
 
     # raise not implemented error if source_Sv corresponds to MVBS
-    if source_Sv.attrs["processing_function"] == "preprocess.compute_MVBS":
+    if source_Sv.attrs["processing_function"] == "commongrid.compute_MVBS":
         raise NotImplementedError("Adding split-beam data to MVBS has not been implemented!")
 
     # check that the appropriate waveform and encode mode have been given
