+++ echopype/preprocess/__init__.py
@@ -1,7 +1,8 @@
-from .api import compute_MVBS, compute_MVBS_index_binning, remove_noise
+from .api import compute_MVBS, compute_MVBS_index_binning, estimate_noise, remove_noise
 
 __all__ = [
+    "estimate_noise",
+    "remove_noise",
     "compute_MVBS",
     "compute_MVBS_index_binning",
-    "remove_noise",
 ]
